PYTHON-1351 Convert cryptography to an optional dependency (#1164)
