+++ requirements.txt
@@ -1,4 +1,3 @@
-cryptography >= 35.0
 geomet>=0.1,<0.3
 six >=1.9
 futures <=2.2.0
