+++ test-datastax-requirements.txt
@@ -1,3 +1,4 @@
 -r test-requirements.txt
 kerberos
 gremlinpython==3.4.6
+cryptography >= 35.0
