+++ tests/integration/standard/test_policies.py
@@ -12,14 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from decimal import Decimal
-import os
-import random
 import unittest
 
 from cassandra.cluster import ExecutionProfile, EXEC_PROFILE_DEFAULT
 from cassandra.policies import HostFilterPolicy, RoundRobinPolicy,  SimpleConvictionPolicy, \
-    WhiteListRoundRobinPolicy, ColDesc, AES256ColumnEncryptionPolicy, AES256_KEY_SIZE_BYTES
+    WhiteListRoundRobinPolicy
 from cassandra.pool import Host
 from cassandra.connection import DefaultEndPoint
 
@@ -27,10 +24,8 @@ from tests.integration import local, use_singledc, TestCluster
 
 from concurrent.futures import wait as wait_futures
 
-
 def setup_module():
     use_singledc()
-
 
 class HostFilterPolicyTests(unittest.TestCase):
 
@@ -93,71 +88,3 @@ class WhiteListRoundRobinPolicyTests(unittest.TestCase):
             queried_hosts.update(response.response_future.attempted_hosts)
         queried_hosts = set(host.address for host in queried_hosts)
         self.assertEqual(queried_hosts, only_connect_hosts)
-
-class ColumnEncryptionPolicyTest(unittest.TestCase):
-
-    def _recreate_keyspace(self, session):
-        session.execute("drop keyspace if exists foo")
-        session.execute("CREATE KEYSPACE foo WITH replication = {'class': 'SimpleStrategy', 'replication_factor': '1'}")
-        session.execute("CREATE TABLE foo.bar(encrypted blob, unencrypted int, primary key(unencrypted))")
-
-    def test_end_to_end_prepared(self):
-
-        # We only currently perform testing on a single type/expected value pair since CLE functionality is essentially
-        # independent of the underlying type.  We intercept data after it's been encoded when it's going out and before it's
-        # encoded when coming back; the actual types of the data involved don't impact us.
-        expected = 12345
-        expected_type = "int"
-
-        key = os.urandom(AES256_KEY_SIZE_BYTES)
-        cl_policy = AES256ColumnEncryptionPolicy()
-        col_desc = ColDesc('foo','bar','encrypted')
-        cl_policy.add_column(col_desc, key, expected_type)
-
-        cluster = TestCluster(column_encryption_policy=cl_policy)
-        session = cluster.connect()
-        self._recreate_keyspace(session)
-
-        prepared = session.prepare("insert into foo.bar (encrypted, unencrypted) values (?,?)")
-        session.execute(prepared, (expected,expected))
-
-        # A straight select from the database will now return the decrypted bits.  We select both encrypted and unencrypted
-        # values here to confirm that we don't interfere with regular processing of unencrypted vals.
-        (encrypted,unencrypted) = session.execute("select encrypted, unencrypted from foo.bar where unencrypted = %s allow filtering", (expected,)).one()
-        self.assertEquals(expected, encrypted)
-        self.assertEquals(expected, unencrypted)
-
-        # Confirm the same behaviour from a subsequent prepared statement as well
-        prepared = session.prepare("select encrypted, unencrypted from foo.bar where unencrypted = ? allow filtering")
-        (encrypted,unencrypted) = session.execute(prepared, [expected]).one()
-        self.assertEquals(expected, encrypted)
-        self.assertEquals(expected, unencrypted)
-
-    def test_end_to_end_simple(self):
-
-        expected = 67890
-        expected_type = "int"
-
-        key = os.urandom(AES256_KEY_SIZE_BYTES)
-        cl_policy = AES256ColumnEncryptionPolicy()
-        col_desc = ColDesc('foo','bar','encrypted')
-        cl_policy.add_column(col_desc, key, expected_type)
-
-        cluster = TestCluster(column_encryption_policy=cl_policy)
-        session = cluster.connect()
-        self._recreate_keyspace(session)
-
-        # Use encode_and_encrypt helper function to populate date
-        session.execute("insert into foo.bar (encrypted, unencrypted) values (%s,%s)",(cl_policy.encode_and_encrypt(col_desc, expected), expected))
-
-        # A straight select from the database will now return the decrypted bits.  We select both encrypted and unencrypted
-        # values here to confirm that we don't interfere with regular processing of unencrypted vals.
-        (encrypted,unencrypted) = session.execute("select encrypted, unencrypted from foo.bar where unencrypted = %s allow filtering", (expected,)).one()
-        self.assertEquals(expected, encrypted)
-        self.assertEquals(expected, unencrypted)
-
-        # Confirm the same behaviour from a subsequent prepared statement as well
-        prepared = session.prepare("select encrypted, unencrypted from foo.bar where unencrypted = ? allow filtering")
-        (encrypted,unencrypted) = session.execute(prepared, [expected]).one()
-        self.assertEquals(expected, encrypted)
-        self.assertEquals(expected, unencrypted)
